move each sale related testsuite into a separate file
